Add the ability to configure searching by extensions
In certain cases searching by extension will be necessary for an
implementation. Due to our dynamic query design that reads off the FHIR
definitions, we are able to support querying by custom extensions with
minimal effort as long as we have certain config values. This change
implements this ideas as it was necesssary for the OpenCRVS
implementation.

OCRVS-910