Allow PUT requests to update or insert records, as required
Prior to this change, requests to the PUT endpoint updated existing
records and returned a 404 response when the requested record did not
exist. This change will prevent having to make two distinct calls
when a requested record isn't found and ensure that the supplied
value for 'id' is assigned to the inserted record.

This change will update an existing record on request; if the record
does not already exist, it will be inserted and response 200
returned.

IHS-121