replace the requests for procedures with observations' requests
Its more realitic that a patient can have two observations in the same encounter than more than one procedure in one encounter. This commit removes all the procedure reequests and it also fixes some bugs. The id that was being passed in as the patient id was actually the history id (same for the encounter id) resulting in the observations request returning nothing. The think function which delays the requests has also been removed beacuse the requests are executed synchronously.

IHS-137