modify the patient encounter and observations create test
The test has been set such that the requests are dependednt on each ohter. An encounter request will be executed only when patient creation is succesful and the observations when the patient and encounter creation is successful. The check done in the function that gets the observations has been changed, it was checking the status but now it checks the number observations created for the patient which should always be two

IHS-137