Pass the logger as an argument to plugins
Pass the initialised logger in as an argument to plugins so that they
can log using the same format and options as the rest of the
application. This is mainly for the sake of external plugins which get
mounted into the plugins folder since they can no longer require Winston
directly.